Add Laravel Passport for API token authentication.
As part of this I removed the default middleware group from the api file
to allow us to use multiple types of authentication in that file.  Please see
the notes in routes/api.php.

This commit requires a few command line arguments to function properly.
1. php artisan migrate
2. php artisan passport:install

One app sign in it should issue a GET to api/user/gettoken with basic authentication.
This will return a long lived (100 year) token.  This token can be revoked in the database but
we don't currently have a ui to do this.  It can be revoked by the app as explained below.

You can retrieve user info with this token by doing a GET to api/v2/me/user using header authentication.

Once the app is done with the token it can do a POST to /api/v2/revoketoken and the current token
will be revoked.